Added post routes file and refactored routes.ts
